fix: use relative paths for includes and flags (#187)
This fix allows RBE to work and have shared cache between systems as no
absolute path cripples in, making all paths in flags deterministic.

A negative side-effect of this fix is that toolchains for all platforms
that are registered will first extract the toolchain files, even for
platforms that are not requested during build.

This negative impact is partially mitigated by PR #186 that
significantly reduces the toolchain archive sizes to about 25% of the
original size.