feat: Add extra_asmflags to pass custom flags to `ACTION_NAME.preprocessor_assemble` (#197)
If found a problem: Some assembly files depend on builtin imports (e.g.
the ones in boringssl). Currently, the way we make those imports
available to targets is by adding all the `-isystem` et.al. flags into
`extra_xxxx` attributes and passing them down to the
cc_toolchain_config. However, the cc_toolchain_config doesn't apply
those flags to `ACTION_NAME.preprocessor_assemble`.

This PR adds an `extra_asmflags` attribute to `gcc_toolchain` (and all
the way down the chain). This solves the issue because we populate that
attribute with the same imports we use for C code, but it also allows
consumers to specify flags for the assembler.

An alternative could be to apply `extra_cflags` to
`ACTION_NAME.preprocessor_assemble`. This is also a good alternative, I
don't have a strong opinion on the matter. @f0rmiga what do you think?